+++ src/flash/nor/lpc2000.c
@@ -1552,7 +1552,7 @@ static int lpc2000_erase_check(struct flash_bank *bank)
 	return lpc2000_iap_blank_check(bank, 0, bank->num_sectors - 1);
 }
 
-static int get_lpc2000_info(struct flash_bank *bank, char *buf, int buf_size)
+static int get_lpc2000_info(struct flash_bank *bank, char *buf, size_t buf_size)
 {
 	struct lpc2000_flash_bank *lpc2000_info = bank->driver_priv;
